fix: move jwk provider types to separate file
