+++ packages/did-provider-jwk/src/types/jwk-provider-types.ts
@@ -1,0 +1,1 @@
+export type JwkDidSupportedKeyTypes = 'Secp256r1' | 'Secp256k1' | 'Ed25519' | 'X25519'
